[BE] refactor: interface는 index로 분리한다.
