+++ BE/src/achievement/dto/achievement-update-request.ts
@@ -1,17 +1,9 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { User } from '../../users/domain/user.domain';
 import { Category } from '../../category/domain/category.domain';
 import { IsNotEmptyString } from '../../config/config/validation-decorator';
-import { IsNotEmpty, IsNumber, IsString } from 'class-validator';
+import { IsNotEmpty, IsNumber } from 'class-validator';
+import { AchievementUpdate } from '../index';
 
-export interface AchievementUpdate {
-  user?: User;
-  category?: Category;
-  title?: string;
-  content?: string;
-  imageUrl?: string;
-  thumbnailUrl?: string;
-}
 export class AchievementUpdateRequest {
   @IsNotEmptyString({ message: '잘못된 제목 이름입니다.' })
   @ApiProperty({ description: 'title' })
@@ -31,7 +23,7 @@ export class AchievementUpdateRequest {
     this.categoryId = categoryId;
   }
 
-  toAchievementUpdate(category: Category) {
+  toAchievementUpdate(category: Category): AchievementUpdate {
     return {
       title: this.title,
       content: this.content,
